requirements.txt
affine @ file:///home/conda/feedstock_root/build_artifacts/affine_1733762038348/work
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
Authlib==1.6.4
autograd==1.8.0
autoray==0.7.2
babel==2.17.0
beautifulsoup4==4.14.0
bleach==6.2.0
blinker==1.9.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1756599151967/work
cachetools==6.2.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1758716116344/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1758270398201/work
click-plugins @ file:///home/conda/feedstock_root/build_artifacts/click-plugins_1750848229740/work
cligj @ file:///home/conda/feedstock_root/build_artifacts/cligj_1733749956636/work
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diastatic-malt==2.15.2
dill==0.4.0
dimod==0.12.21
diskcache==5.6.3
dwave-cloud-client==0.14.0
dwave-gate==0.3.4
dwave-hybrid==0.6.14
dwave-inspector==0.5.5
dwave-neal==0.6.0
dwave-ocean-sdk==9.0.0
dwave-optimization==0.6.4
dwave-preprocessing==0.6.10
dwave-samplers==1.6.0
dwave-system==1.33.0
dwave_networkx==0.8.18
executing==2.2.1
fasteners==0.20
fastjsonschema==2.21.2
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1755216263872/work
Flask==3.1.2
fonttools==4.60.0
fqdn==1.5.1
gast==0.6.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1756739574751/work
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
homebase==1.0.1
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
http-sf==1.0.4
httpcore==1.0.9
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.3
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.3.0
lazy_loader==0.4
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
matplotlib==3.10.6
matplotlib-inline==0.1.7
minorminer==0.2.19
mistune==3.1.4
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1749078300/work
notebook_shim==0.2.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225376651/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=d08e1c9e5833ae7780563812aa73e2497db1ee3bd5510d3becb8aa18aa2d0c7c
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
penaltymodel==1.3.0
pennylane==0.42.3
pennylane_lightning==0.42.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plucky==0.4.3
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1758436411/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pyspnego==0.12.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
pyzmq==27.1.0
qiskit==2.2.1
qiskit-aer==0.17.2
qiskit-ibm-runtime==0.42.0
-e git+ssh://git@github.com/ailabteam/quantum-sar.git@c449dc947bd4a25a8633c9c9af5c0789083a5f9b#egg=quantum_sar
rasterio==1.4.3
referencing==0.36.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
requests_ntlm==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rustworkx==0.17.1
scikit-image==0.25.2
scipy==1.16.2
scipy-openblas32==0.3.30.0.2
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
snuggs @ file:///home/conda/feedstock_root/build_artifacts/snuggs_1733818638588/work
soupsieve==2.8
stack-data==0.6.3
stevedore==5.5.0
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
termcolor==3.1.0
terminado==0.18.1
tifffile==2025.9.20
tinycss2==1.4.0
tomlkit==0.13.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
zipp==3.23.0
zstandard==0.23.0
